Reorganise and refactor handling on header nesting
This commit reorganises the `docs` module into a couple of submodules
for easier navigation. It also replaces manual specifying of nesting
level upon Markdown element creation for automatic header generation.

I should note here that I've decided to simplify the code over the
previous commit, as the previous seemed unnecessary bloated and
really subtle in places which IMHO would very likely cause some
maintenance hell in the future.
